2024/1/13 1:25PM
completely rehauled the structure of how this project works.

all of the old basic methods remain unchanged, the tester, deck, and cards. I added a GameActor file that contains a game actor and dealer class. the game actor will contain all of the data of a player, and if a player splits the data the split hand.

The constructor of the game_actions class is now split into 2 main constructor logic flows. If the number of games is specified, we assume this is a user who wants to play multiple hands. If it is not specified, we assume that we want a data-getting game to be played (no screen printing or input for actions, just logic dependencies)