2024/1/14 5:40PM
I just realized I tried committing this twice to the portfolio. Midflight commit bless up delta free wifi

Added the splitting option, added the way for the winner to be determined.

deciding to make the class be its own game class and then have a seperate class for the data games.

Just a few more quick fixes and we should be off to the races.